fix jira：PLATFORM-1244
The difference between web app root and servlet application contexts is
one of the top sources of confusion in Spring, see difference between
applicationContext and spring-servlet.xml in spring
From @Value javadoc :
Note that actual processing of the @Value annotation is performed by a
BeanPostProcessor
From Spring documentation:
BeanPostProcessor interfaces are scoped per-container. This is only
relevant if you are using container hierarchies. If you define a
BeanPostProcessor in one container, it will only do its work on the
beans in that container. Beans that are defined in one container are not
post-processed by a BeanPostProcessor in another container, even if both
containers are part of the same hierarchy.
resole method：
add properties's config to the dispatcher-servlet.xml
!!!!!!: This also works for injection of property values into Spring MVC
interceptors as well, e.g., subclasses of HandlerInterceptorAdaptor